feat: componentes criados

+++ frontend/aurore/src/app/components/card/card.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnInit} from '@angular/core';
 import { Router } from '@angular/router';
 import { ManegementService } from 'src/app/services/management/management.service';
-
+import { Location } from '@angular/common';
 @Component({
   selector: 'app-card',
   templateUrl: './card.component.html',
@@ -13,32 +13,37 @@ export class CardComponent {
   @Input() rotaDel: any;
   rotaEdit: any;
 
-  constructor(private router: Router, private service: ManegementService){}
+  constructor(private router: Router, private service: ManegementService, private location: Location){}
 
   rotaChange(): void {
     if (this.reserva && this.rota) {
       const rotaEdit = `${this.rota}/${this.reserva.id}`;
-      console.log("rota ID",rotaEdit)
-      this.router.navigateByUrl(rotaEdit);
+  
+      const dados = {
+        accommodation: this.reserva.accommodation_id
+      };
+
+      this.router.navigate([rotaEdit], { state: { dados } });
     }
   }
 
   deleteCard(){
     if(this.reserva && this.rotaDel){
-      const url = `${this.rotaDel}/${this.reserva.id}/delete`
 
-      console.log("URL FINAL",url);
-      this.service.deleteId(url).subscribe((result)=>{
-        console.log("resultado del:", result)
-        if(result.status_code == 200) {
-          alert('Reserva deletada com sucesso!');
-          setTimeout(function() {
-            location.reload();
-          }, 1000);
+      const confirmacao = window.confirm("Você deseja deletar esta moradia?");
 
-        }
-      });
+      if(confirmacao){
 
+        const url = `${this.rotaDel}/${this.reserva.id}/delete`
+
+        this.service.deleteId(url).subscribe((result)=>{
+          alert(result.detail);
+          window.location.reload();
+          
+        });
+
+      }
+     
     }
   }
